backend/requirements.txt
# Core ML/DL libraries
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0

# Computer Vision
opencv-python>=4.8.0
mediapipe>=0.10.0
Pillow>=9.5.0

# Data Processing
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0

# Web Framework & Real-time Communication
fastapi==0.104.1
uvicorn>=0.24.0
websockets>=12.0
python-multipart>=0.0.6
pydantic==2.0.3
pydantic-settings==2.0.3
python-jose[cryptography]>=3.3.0

# Development
pytest>=7.4.0
black>=23.0.0